:bug: [peer] Fixed freeze in peer.repair().
Fixed a case in peer.repair() where a freshly dialed
connection was left to rot, which caused the connection to
hang.
Changed the receiver and subscription interfaces to be more
usable. Receiver can now detect whether it is subscribed to
anything, and can be used to abort waiting for messages in
case it is not subscribed to any peers.

Added test for Peer.Close().