:recycle: [peer] Fixed edge case in Mutex.TryLock(), refactored Peer.Send().
Now, Mutex.TryLock() will always fail if called with a
context that is already canceled. To have a non-blocking
TryLock(), a nil context has to be passed. Added a test that
ensures that a canceled context will never lock a mutex.
Made Peer.Send()'s initial abort condition more readable.